refactor: new avc design (#397)
* refactor: developerActInflTrend

* refactor: repoActiInflTrend

* refactor: change directories

* refactor: developerNetwork

* refactor: perceptortab

* refactor: perceptorLayout

* refactor: projectNetwork

* refactor: teachingbubble

* refactor: hypertrons

* refactor:perceptor

* refactor: divide perceptorLayout

* refactor: since we import css files as modules there is no need to load them via manifest

* refactor: renameing & restructure

* fix: contentscript cannot auto-reload after code structure changes
