+++ src/views/DeveloperActiInflTrendView/DeveloperActiInflTrendView.tsx
@@ -1,13 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { render } from 'react-dom';
-import $ from 'jquery';
-import * as pageDetect from 'github-url-detection';
-import {
-  getGithubTheme,
-  runsWhen,
-  getMessageByLocale,
-} from '../../utils/utils';
-import PerceptorBase from './PerceptorBase';
+import { getGithubTheme, getMessageByLocale } from '../../utils/utils';
 import Settings, { loadSettings } from '../../utils/settings';
 import { getDeveloperActiInfl } from '../../api/developer';
 import Bars from '../../components/Bars/index';
@@ -100,46 +92,4 @@ const DeveloperActiInflTrendView: React.FC<DeveloperActiInflTrendViewProps> = ({
   );
 };
 
-@runsWhen([pageDetect.isUserProfile])
-class DeveloperActiInflTrend extends PerceptorBase {
-  private _currentDeveloper: string;
-
-  constructor() {
-    super();
-    this._currentDeveloper = '';
-  }
-
-  public async run(): Promise<void> {
-    this._currentDeveloper = $('.p-nickname.vcard-username.d-block')
-      .text()
-      .trim();
-
-    let profileArea = null;
-    let newContainer = null;
-
-    // if exists (when going backword or forward in browser history)
-    if (document.getElementById('developer-acti-infl-trend') != null) {
-      render(
-        <DeveloperActiInflTrendView
-          currentDeveloper={this._currentDeveloper}
-        />,
-        document.getElementById('developer-acti-infl-trend')
-      );
-    } else {
-      profileArea = $('.js-profile-editable-area').parent();
-      newContainer = document.createElement('div');
-      newContainer.id = 'developer-acti-infl-trend';
-      newContainer.style.width = '100%';
-
-      render(
-        <DeveloperActiInflTrendView
-          currentDeveloper={this._currentDeveloper}
-        />,
-        newContainer
-      );
-      profileArea.after(newContainer);
-    }
-  }
-}
-
-export default DeveloperActiInflTrend;
+export default DeveloperActiInflTrendView;
