+++ src/views/PerceptorLayoutView/PerceptorLayoutView.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PerceptorLayoutView: React.FC = () => {
+  return <div />;
+};
+
+export default PerceptorLayoutView;
