+++ src/views/TeachingBubbleWrapperView/TeachingBubbleWrapperView.tsx
@@ -9,7 +9,9 @@ export interface TeachingBubbleProps {
   target: string;
 }
 
-const TeachingBubbleWrapper: React.FC<TeachingBubbleProps> = ({ target }) => {
+const TeachingBubbleWrapperView: React.FC<TeachingBubbleProps> = ({
+  target,
+}) => {
   const [metaData, setMetaData] = useState(new MetaData());
   const [inited, setInited] = useState(false);
   const [settings, setSettings] = useState(new Settings());
@@ -75,4 +77,4 @@ const TeachingBubbleWrapper: React.FC<TeachingBubbleProps> = ({ target }) => {
   );
 };
 
-export default TeachingBubbleWrapper;
+export default TeachingBubbleWrapperView;
